Addressed PR comments and added functionality to TimeCalculator to calculate days worked and weeks worked. Also added svg icon on permanent employee page for delete button in the table

+++ PayCalculator/PayCalculatorMVC/Controllers/PermanentEmployeeController.cs
@@ -84,6 +84,11 @@ namespace PayCalculatorMVC.Controllers
         [HttpPost]
         public IActionResult Deletion(int id)
         {
+            if (!_permEmployeeRepo.Delete(id))
+            {
+                return RedirectToAction("Index");
+            }
+
             _permEmployeeRepo.Delete(id);
             return RedirectToAction("Index");
         }
